requirements.txt
browserstack-local
setuptools
selenium
requests
Appium-Python-Client
